+++ components/Collection.js
@@ -21,7 +21,7 @@ function Collection({ id, term }) {
 
   let description = collection.description || ""
   if (highlight) {
-    const searchMask = `(${highlight})`
+    const searchMask = `(${highlight.replaceAll(" ", "|")})`
     const regEx = new RegExp(searchMask, "ig")
     const replaceMask = "<mark>$1</mark>"
     description = description.replace(regEx, replaceMask)
@@ -35,7 +35,7 @@ function Collection({ id, term }) {
         </h2>
       </Link>
       <div className="text-xs">
-        Agency:{` `}
+        Data Supply Agency:{` `}
         <Link href={`/agencies/${collection.agency.agency_id}`}>
           <a className="underline cursor-pointer">
             {collection.agency.agency_name}