+++ components/Search.js
@@ -2,7 +2,6 @@ import { CogIcon, FilterIcon } from "@heroicons/react/outline"
 import { usePlausible } from "next-plausible"
 import { useRouter } from "next/router"
 import { useEffect, useRef, useState } from "react"
-import { event } from "../lib/gtag"
 
 function Search() {
   const router = useRouter()
@@ -13,8 +12,6 @@ function Search() {
   const [showConfig, setShowConfig] = useState(false)
   const [searchRefreshes, setSearchRefreshes] = useState(true)
   const [searchAdhoc, setSearchAdhoc] = useState(true)
-  const [searchMetadata, setSearchMetadata] = useState(true)
-  const [searchRnD, setSearchRnD] = useState(true)
 
   const plausible = usePlausible()
 
@@ -30,19 +27,16 @@ function Search() {
     }
     if (value === "") q = rest
     if (value !== "") {
-      event("search", "general", value)
       plausible("Search", {
         props: {
           term: value,
         },
       })
     }
-    if (!(searchRefreshes && searchAdhoc && searchMetadata && searchRnD)) {
+    if (!(searchRefreshes && searchAdhoc)) {
       let include = []
       if (searchRefreshes) include.push("refreshes")
       if (searchAdhoc) include.push("adhoc")
-      if (searchMetadata) include.push("meta")
-      if (searchRnD) include.push("rnd")
       q.include = include.join("_")
     }
     router.push(
@@ -123,34 +117,18 @@ function Search() {
               />{" "}
               Adhoc
             </div>
-
-            {/* <div className="flex gap-2 items-center">
-              <input
-                type="checkbox"
-                checked={searchMetadata}
-                onChange={(e) => setSearchMetadata(e.target.checked)}
-              />{" "}
-              Metadata
-            </div> */}
-
-            <div className="flex gap-2 items-center">
-              <input
-                type="checkbox"
-                checked={searchRnD}
-                onChange={(e) => setSearchRnD(e.target.checked)}
-              />{" "}
-              RnD
-            </div>
           </div>
         </div>
       )}
 
-      <div className="flex gap-2 justify-end items-center text-xxs mx-2">
-        <div className="flex items-center gap-1">
-          <span className="h-2 w-2 rounded-full bg-green-400"></span>
-          <span>Metadata available</span>
+      {router.pathname !== "/agencies" && (
+        <div className="flex gap-2 justify-end items-center text-xxs mx-2">
+          <div className="flex items-center gap-1">
+            <span className="h-2 w-2 rounded-full bg-green-400"></span>
+            <span>Metadata available</span>
+          </div>
         </div>
-      </div>
+      )}
     </form>
   )
 }