+++ changelog.md
@@ -3,10 +3,12 @@
 **Updated 19 October 2022**
 
 - Tweaked colours and text on the 'Quick Stats' page
-- Added October 2022 refresh data
+- Added October 2022 refresh data and removed _old_ adhoc variables
+- merged most duplicate collections
 - reverse 'refresh availability' so newest first
 - use SQL fulltext search for better search results
 - update about and help pages
+- specify 'Data Supply Agency'
 
 # version 0.2.0